test_instance: move LXD-only test to LXD-specific test path
And remove the parameterisation, as we only have a single parameter set.
