+++ pycloudlib/tests/test_instance.py
@@ -4,7 +4,6 @@ from unittest import mock
 import pytest
 
 from pycloudlib.instance import BaseInstance
-from pycloudlib.lxd.instance import LXDInstance
 from pycloudlib.result import Result
 
 # Disable this pylint check as fixture usage incorrectly triggers it:
@@ -19,36 +18,6 @@ def concrete_instance_cls():
     """
     with mock.patch.object(BaseInstance, "__abstractmethods__", set()):
         yield BaseInstance
-
-
-class TestExecute:
-    """Tests covering pycloudlib.instance.Instance.execute.
-
-    TODO: There are elements of `execute` which could be refactored onto the
-          relevant subclasses.  Some of these tests should move along with that
-          refactor.
-    """
-
-    @pytest.mark.parametrize(
-        "instance_cls,cloud_name",
-        (
-            (LXDInstance, "lxd"),
-        ),
-    )
-    def test_all_rcs_acceptable(self, instance_cls, cloud_name):
-        """Test that we invoke util.subp with rcs=None.
-
-        rcs=None means that we will get a Result object back for all return
-        codes, rather than an exception for non-zero return codes.
-        """
-        instance = instance_cls(None)
-        with mock.patch(
-            "pycloudlib.{}.instance.subp".format(cloud_name)
-        ) as m_subp:
-            instance.execute("some_command")
-        assert 1 == m_subp.call_count
-        _args, kwargs = m_subp.call_args
-        assert kwargs.get("rcs", mock.sentinel.not_none) is None
 
 
 class TestWait:
